# merge and join speedups