This is a rewrite of `merge` and `join` to improve performance by an order of magnitude.

```julia
using AxisArrays
data = randn(400, 100, 50);
A1 = AxisArray(data, Axis{:time}(1:400), Axis{:lon}(1:100), Axis{:lat}(1:50))
A2 = AxisArray(data, Axis{:time}(401:800), Axis{:lon}(1:100), Axis{:lat}(1:50))
```

Master:

```julia
julia> @time merge(A1, A2);
 11.517995 seconds (113.78 M allocations: 4.252 GB, 4.12% gc time)

julia> @time join(A1, A2);
 15.268915 seconds (131.22 M allocations: 5.025 GB, 4.67% gc time)
```

This PR:

```julia
julia> @time merge(A1, A2);
  0.806035 seconds (16.01 M allocations: 1.788 GB, 23.98% gc time)

julia> @time join(A1, A2);
  0.838161 seconds (16.01 M allocations: 1.848 GB, 23.06% gc time)
```