@timholy 2017-04-06
16M allocations still seems awfully suspicious. Have you looked into `@profile`/ProfileView and `@code_warntype`?