@timholy 2017-04-06
With a simple change I can get another factor of 10:
```julia
julia> @time merge(A1, A2);
  0.091910 seconds (5.30 k allocations: 61.273 MiB, 74.16% gc time)
```

The idea was to delete the conversion to a scalar index and just use the original cartesian indexing:
```diff
$ git diff
diff --git a/src/combine.jl b/src/combine.jl
index 5ac765a..4c42a27 100644
--- a/src/combine.jl
+++ b/src/combine.jl
@@ -53,13 +53,9 @@ function scalarinds{N}(dims::NTuple{N,Int}, inds::NTuple{N,Vector{Int}})
 end
 
 function indexmappings{N}(oldaxes::NTuple{N,Axis}, newaxes::NTuple{N,Axis})
-
     oldvals = axisvalues(oldaxes...)
     newvals = axisvalues(newaxes...)
     before_idxs, after_idxs = zip(indexmapping.(oldvals, newvals)...) |> collect
-
-    return scalarinds(length.(oldaxes), before_idxs), scalarinds(length.(newaxes), after_idxs)
-
 end
 
 function indexmapping(old::AbstractVector, new::AbstractVector)
@@ -113,7 +109,7 @@ function Base.merge{T,N,D,Ax}(As::AxisArray{T,N,D,Ax}...; fillvalue::T=zero(T))
 
     for A in As
         before_idxs, after_idxs = indexmappings(A.axes, result.axes)
-        result.data[after_idxs] = A.data[before_idxs]
+        result.data[after_idxs...] = A.data[before_idxs...]
     end
 
     return result
```
It's still got some type-instability problems, though. (`@inferred(merge(A1, A2))` fails.)