@GordStephen 2017-04-06
Awesome! I though I'd tried direct indexing before writing `scalarinds` but couldn't get it to work for some reason, looking at it now it seems like the obvious solution though. I've been using `@profile` but am not familiar with `@code_warntype`, I'll look into that more.

Interesting about the type stability, it's not immediately clear to me what the reason for that is. With `merge` in most cases the output type should match the input types (the only exception being range axes which are always converted to vectors - likely another performance improvement opportunity, but from a type-stability perspective at least it's consistent?) I'll read up on `@inferred` and dig into that a bit more.