@timholy 2017-04-06
I suspect it's
```julia
resultaxes = map(as -> axismerge(:outer, as...), zip(axes.(As)...))
```
You'll have to keep everything in the tuple-world (rather than Array-world) if you want this to be type-stable.