@GordStephen 2017-04-08
Ok, `merge` and `join` appear to be type-stable now. Interestingly enough, it's only the `outer` case of `axismerge` that drives the requirement to convert range axes to vectors, so it's probably worth looking at something like PiecewiseIncreasingRanges.jl as a lighter-weight alternative to vector conversion.

Can the `AbstractArray` argument to `_nextaxistype` be eliminated? The generated function [doesn't seem to use](https://github.com/JuliaArrays/AxisArrays.jl/blob/v0.1.1/src/core.jl#L171) its value or element type parameter?