@timholy 2017-04-08
Seems fine to me. Can you rebase on current master? 0.6 is failing its tests, presumably will be fixed by a rebase that includes #74.